Version 1.0

+++ DragAndDropFluid/Model/Point2D.cs
@@ -1,9 +1,9 @@
 ﻿namespace DragAndDropFluid.Model
 {
-	public class Point2d(float x, float y)
+	public class Point2d(double x, double y)
 	{
-		public float X { get; set; } = x;
-		public float Y { get; set; } = y;
+		public double X { get; set; } = x;
+		public double Y { get; set; } = y;
 
 		public override string ToString()
 		{
